+++ x/superfluid/keeper/epoch.go
@@ -7,6 +7,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	distributiontypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
 	"github.com/osmosis-labs/osmosis/osmoutils"
 	cl "github.com/osmosis-labs/osmosis/v26/x/concentrated-liquidity"
@@ -77,6 +78,11 @@ func (k Keeper) MoveSuperfluidDelegationRewardToGauges(ctx sdk.Context, accs []t
 		_ = osmoutils.ApplyFuncIfNoError(ctx, func(cacheCtx sdk.Context) error {
 			_, err := k.ck.WithdrawDelegationRewards(cacheCtx, addr, valAddr)
 			if errors.Is(err, distributiontypes.ErrEmptyDelegationDistInfo) {
+				ctx.Logger().Debug("no delegations for this (pool, validator) pair, skipping...")
+				// TODO: Remove this account from IntermediaryAccounts that we iterate over
+				return nil
+			} else if errors.Is(err, stakingtypes.ErrNoDelegation) {
+				// NOTE: in v0.50.x the function changed to return a different error
 				ctx.Logger().Debug("no delegations for this (pool, validator) pair, skipping...")
 				// TODO: Remove this account from IntermediaryAccounts that we iterate over
 				return nil